requirements.txt
smof
biopython
click
tqdm
